swap types in numeric fn

+++ source/numeric.hpp
@@ -134,7 +134,7 @@ template <typename T> T interpolate(const T& a, const T& b, Float t)
 }
 
 
-inline u8 fast_interpolate(u32 a, u32 b, u32 t)
+inline u8 fast_interpolate(u8 a, u8 b, u8 t)
 {
     return b + (u8)(((u16)(a - b) * t) >> 8);
 }
